docs: rename registry to templates
