+++ apps/docs/app/[lang]/resources/page.tsx
@@ -1,5 +1,5 @@
 import { permanentRedirect } from "next/navigation";
 
-const ResourcesPage = () => permanentRedirect("/registry");
+const ResourcesPage = () => permanentRedirect("/templates");
 
 export default ResourcesPage;
