+++ apps/docs/app/[lang]/templates/[slug]/file-viewer.tsx
